feat: added buttons map

+++ app/src/main/java/ru/gsw/way_to_success_ii/buttons/gameButtons/GameButtonParameters.kt
@@ -2,6 +2,6 @@ package ru.gsw.way_to_success_ii.buttons.gameButtons
 
 class GameButtonParameters {
     var title: String = ""
-    var price: Int = 0
+    var price: Long = 0
     var onClick = {}
 }
